updated profile.js
made reusable components
